convert editorial display
