+++ src/client/apps/edit/components/display/index.styl
@@ -1,35 +1,3 @@
-.EditDisplay
-  padding 40px 0
-  .DropDownItem__content
-    margin-top 40px
-  .field-group.flat-checkbox
-    display flex
-    align-items center
-  .edit-display__preview
-    margin-left 20px
-  .title-fill
-    color gray-dark-color
-    garamond l-caption
-    text-align right
-    width 100%
-    margin-top 5px
-    cursor pointer
-
-.DisplayEmail
-  .col--right
-    display flex
-    flex-direction column
-    justify-content space-between
-
-  .field-group--inline
-    margin-left 20px
-    .field-group
-      width 100%
-      &:first-child
-        margin-right 15px
-      &:last-child
-        margin-left 15px
-
 .DisplayPartner
   &__preview
     background-color gray-lightest-color
